pyproject.toml
[project]
name = "repl-nix-malebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "yt-dlp>=2024.10.22",
]
